moved scoring types to sub-package
